pyproject.toml
[tool.poetry]
name = "polars_helper"
version = "0.1.0"
description = "Package to help developing and testing in Polars"
readme = "README.md"
authors = ["Georgio"]
packages = [
    {include = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "2.2.1"
polars-lts-cpu = "0.20.16"
pyarrow = "15.0.2"
setuptools = "^69.2.0"

[tool.poetry.dev-dependencies]
black = "24.3.0"
pytest = "8.1.1"
flake8 = "7.0.0"

[build-system]
requires = ["poetry-core", "setuptools>=40.8.0"]


